refactor: split consent controller into send and pull
